Changed the numbering system.
Now pretty much everything is numbering in a linear way. The only exception are claims, which there are now two kinds of. Unnumbered claimes (claim* enviroment) behave exactly as they name indicates. Numbered claims (claim enviroment) are numbered, but they are intended to only be used locally in proof enviroments; in particular the counter resets afterwards.